add: update button

+++ main.py
@@ -18,9 +18,10 @@ def markup(lang:str='en') -> telebot.types.InlineKeyboardMarkup:
         m.add(telebot.types.InlineKeyboardButton('我也试试', switch_inline_query_current_chat=''), telebot.types.InlineKeyboardButton('创建一个待办', url='t.me/arielstodolistsbot'))
     return m
 
-def closeMarkup(from_message: telebot.types.Message) -> telebot.types.InlineKeyboardMarkup:
+def closeMarkup(listNamesArg: str, chatID: int, messageID: int, userID: int) -> telebot.types.InlineKeyboardMarkup:
     m = telebot.types.InlineKeyboardMarkup()
-    m.add(telebot.types.InlineKeyboardButton('Close', callback_data=f'{from_message.chat.id} {from_message.message_id}'))
+    m.add(telebot.types.InlineKeyboardButton('Update', callback_data=f'{chatID} {messageID} {userID} update {listNamesArg}'),
+            telebot.types.InlineKeyboardButton('Close', callback_data=f'{chatID} {messageID} {userID} close'))
     return m
 
 @bot.message_handler()
@@ -51,7 +52,7 @@ async def reply(message: telebot.types.Message) -> int:
         elif cmd == '/get':
             flag = True
             f = True
-            r = await bot.reply_to(message, event.get(message.from_user.id, arg), parse_mode='html', reply_markup=closeMarkup(message))
+            r = await bot.reply_to(message, event.get(message.from_user.id, arg), parse_mode='html', reply_markup=closeMarkup(arg, message.chat.id, message.message_id, message.from_user.id))
         elif cmd == '/mark':
             r = await bot.reply_to(message, event.mark(message.from_user.id, arg), parse_mode='html')
         elif cmd == '/add':
@@ -89,10 +90,13 @@ async def inline_reply(inline_query: telebot.types.InlineQuery):
 @bot.callback_query_handler(lambda _: True)
 async def delete_on_callback(callback_query: telebot.types.CallbackQuery):
     try:
-        chatID1, messageID1 = callback_query.data.split(' ')
+        chatID1, messageID1, userID, operation = callback_query.data.split(' ', 3)
         chatID2, messageID2 = callback_query.message.chat.id, callback_query.message.message_id
-        await bot.delete_message(int(chatID1), int(messageID1))
-        await bot.delete_message(int(chatID2), int(messageID2))
+        if operation == 'close':
+            await bot.delete_message(int(chatID1), int(messageID1))
+            await bot.delete_message(int(chatID2), int(messageID2))
+        elif operation.startswith('update '):
+            await bot.edit_message_text(event.get(userID, operation[7:]), chatID2, messageID2, parse_mode='html', reply_markup=closeMarkup(operation[7:], int(chatID1), int(messageID1), int(userID)))
     except Exception as e:
         print(f'Error when handling a "close" request: {e}')
